+++ src/main/kotlin/calculator/CalculatorValidator.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 object CalculatorValidator {
