+++ src/test/kotlin/calculator/StringCalculatorTest.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.assertThat
